requirements.txt
pandas
pandasql
openai
python-dotenv
flask
openpyxl
